+++ src/hooks/service/authHooks.ts
@@ -1,5 +1,7 @@
 import { useMutation } from '@tanstack/react-query';
 import { login, type LoginResponse } from '@/apis/auth';
+import { useContext } from 'react';
+import { AuthContext } from '@/contexts/AuthContext';
 
 interface LoginVariables {
   email: string;
@@ -11,3 +13,9 @@ export function useLoginMutation() {
     mutationFn: login,
   });
 }
+
+export const useAuth = () => {
+  const context = useContext(AuthContext);
+  if (!context) throw new Error('useAuth must be used within AuthProvider');
+  return context;
+};
