--- src/hooks/useAuth.ts
@@ -1,8 +1,0 @@
-import { useContext } from 'react';
-import { AuthContext } from '@/contexts/AuthContext';
-
-export const useAuth = () => {
-  const context = useContext(AuthContext);
-  if (!context) throw new Error('useAuth must be used within AuthProvider');
-  return context;
-};
