--- src/hooks/useThemesQuery.ts
@@ -1,9 +1,0 @@
-import { useSuspenseQuery } from '@tanstack/react-query';
-import { fetchThemes, type Theme } from '@/apis/fetchThemes';
-
-export function useThemesQuery() {
-  return useSuspenseQuery<Theme[]>({
-    queryKey: ['themes'],
-    queryFn: fetchThemes,
-  });
-}
