+++ src/pages/OrderPage.tsx
@@ -15,9 +15,8 @@ import { orderSchema, type OrderFormData } from '@/utils/validateOrderSchema';
 import * as z from 'zod';
 import { toast } from 'react-toastify';
 import { AxiosError, HttpStatusCode } from 'axios';
-import { useAuth } from '@/hooks/useAuth';
-import { useOrderMutation } from '@/hooks/useOrderMutation';
-import { useProductSummaryQuery } from '@/hooks/useProductSummaryQuery';
+import { useAuth } from '@/hooks/service/authHooks';
+import { useOrderMutation, useProductSummaryQuery } from '@/hooks/queries/orderHooks';
 import { Suspense } from 'react';
 import Spinner from '@/components/Spinner';
 import { ErrorBoundary } from 'react-error-boundary';
