+++ src/sections/ThemeSection/ThemeProductSection.tsx
@@ -2,7 +2,7 @@ import styled from '@emotion/styled';
 import { useEffect, useRef } from 'react';
 import { useNavigate } from 'react-router';
 import Spinner from '@/components/Spinner';
-import { useThemeProductsQuery } from '@/hooks/useThemeProductsQuery';
+import { useThemeProductsQuery } from '@/hooks/queries/useThemesQuery';
 import { type ThemeProduct } from '@/apis/theme';
 
 interface ThemeProductSectionProps {
